requirements.txt
flask>=1.1.0
google-cloud-storage
gunicorn==20.0.4